Fix accidental dev invocation. (#803)
Dev mode ignores single instancing. However, whether we launch in dev mode is checked based on the presence of a .gitignore file *in the current path*. This check was too simplistic and would result in accidental dev invocations.

For example, when the user has a global .gitignore file in their $HOME directory. Launchers, e.g. rofi, will use $HOME as their current directory. The simplistic check will assume the global .gitignore file to be this project's file instead.